1.[MODIFY]:modify vector3f to vector3i

+++ include/Triangle.hpp
@@ -15,7 +15,7 @@ struct Triangle {
 
   void setVertex(std::initializer_list<Eigen::Vector3f> _vertex);
   void setNormal(std::initializer_list<Eigen::Vector3f> _normal);
-  void setColor(std::initializer_list<Eigen::Vector3f> _color);
+  void setColor(std::initializer_list<Eigen::Vector3i> _color);
   void setTexCoord(std::initializer_list<Eigen::Vector2f> _texCoords);
 
   /*the original coordinates of the triangle, v0, v1, v2 in counter clockwise
@@ -23,7 +23,7 @@ struct Triangle {
   std::array<Eigen::Vector3f, 3> m_vertex;
 
   // Color for each vertex
-  std::array<Eigen::Vector3f, 3> m_color;
+  std::array<Eigen::Vector3i, 3> m_color;
 
   // texture u,v coordinates for each vertex
   std::array<Eigen::Vector2f, 3> m_texCoords;
